pyproject.toml
[project]
name = "videofastapi"
version = "0.1.0"
description = "Subscription-based file/video hosting for a single author"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=4.2.0",
    "email-validator>=2.2.0",
    "fastapi>=0.128.8",
    "gigachat>=0.2.0",
    "imagekitio>=5.4.0",
    "pydantic[email]>=2.9.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.12",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn[standard]>=0.39.0",
]
